+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/dialog/DefaultLoginDialogBuilder.java
@@ -73,7 +73,8 @@ final class DefaultLoginDialogBuilder extends AbstractDialogBuilder<LoginDialogB
 			owner = dummyFrame = createDummyFrame(title == null ? null : title.get(), icon);
 		}
 		LoginPanel loginPanel = new LoginPanel(defaultUser, validator, icon, southComponent, inputFieldColumns);
-		OkCancelDialogBuilder dialogBuilder = new DefaultOkCancelDialogBuilder(loginPanel)
+		OkCancelDialogBuilder dialogBuilder = new DefaultOkCancelDialogBuilder()
+						.component(loginPanel)
 						.owner(owner)
 						.resizable(false)
 						.title(title)